+++ src/runtimes/node/bundlers/nft/index.ts
@@ -9,7 +9,7 @@ import { cachedReadFile, FsCache } from '../../../../utils/fs.js'
 import { minimatch } from '../../../../utils/matching.js'
 import { getBasePath } from '../../utils/base_path.js'
 import { filterExcludedPaths, getPathsOfIncludedFiles } from '../../utils/included_files.js'
-import type { GetSrcFilesFunction, BundleFunction } from '../index.js'
+import type { GetSrcFilesFunction, BundleFunction } from '../types.js'
 
 import { processESM } from './es_modules.js'
